Factored out UI bootloader module
- Factored out bootloader main function
- Factored out handlers for bolos ux ids
- Added "bootloader mode" to bootloader main which hangs execution and never returns after onboard
