+++ ledger/src/ui/src/bolos_ux.h
@@ -208,9 +208,6 @@ unsigned int bolos_ux_get_word_ptr(unsigned char **word,
  */
 void bolos_ux_main(void);
 
-// Check if app is allowed, and manage app blacklist
-int is_app_version_allowed(application_t *app);
-
 void screen_display_init(unsigned int stack_slot);
 void screen_state_init(unsigned int stack_slot);
